DataBase is added & App Refactored
