+++ .idea/misc.xml
@@ -8,6 +8,7 @@
         <entry key="..\:/Users/Moham/Desktop/MyApplication/app/src/main/res/drawable/card_back_male_female.xml" value="0.2445" />
         <entry key="..\:/Users/Moham/Desktop/MyApplication/app/src/main/res/drawable/card_back_plus_minus.xml" value="0.2445" />
         <entry key="..\:/Users/Moham/Desktop/MyApplication/app/src/main/res/drawable/ic_baseline_info_24.xml" value="0.121" />
+        <entry key="..\:/Users/Moham/Desktop/MyApplication/app/src/main/res/layout/activity_bmi_page.xml" value="0.25" />
         <entry key="..\:/Users/Moham/Desktop/MyApplication/app/src/main/res/layout/activity_bmi_result.xml" value="0.264" />
         <entry key="..\:/Users/Moham/Desktop/MyApplication/app/src/main/res/layout/activity_main.xml" value="0.4" />
       </map>
