+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApplication">
         <activity
+            android:name=".BMI_Page"
+            android:exported="false" />
+        <activity
             android:name=".BMI_Result"
             android:exported="false" />
         <activity
